Add instance.mountWithHidden
