+++ src/index.js
@@ -1,12 +1,12 @@
+// @flow
 import createShut from './createShut.jsx'
 import {
+  seedFromTop,
   seedFromBottom,
   seedFromLeft,
-  seedFromRight,
-  seedFromTop
-} from './seeds'
-
+  seedFromRight
+} from './seeds.js'
+export const ShutFromTop = createShut(seedFromTop)
 export const ShutFromBottom = createShut(seedFromBottom)
 export const ShutFromLeft = createShut(seedFromLeft)
 export const ShutFromRight = createShut(seedFromRight)
-export const ShutFromTop = createShut(seedFromTop)
