--- src/seeds/index.js
@@ -1,6 +1,0 @@
-import seedFromBottom from './fromBottom'
-import seedFromLeft from './fromLeft'
-import seedFromRight from './fromRight'
-import seedFromTop from './fromTop'
-
-export { seedFromBottom, seedFromLeft, seedFromRight, seedFromTop }
