+++ test/seeds.js
@@ -5,7 +5,7 @@ import {
   seedFromLeft,
   seedFromRight,
   seedFromTop
-} from '../src/seeds'
+} from '../src/seeds.js'
 
 describe(`firstRootSize`, () => {
   const createReact = () => ({})
@@ -127,7 +127,7 @@ describe(`canInit`, () => {
   })
 })
 
-describe(`onTouchMove`, () => {
+describe(`onTouchMoveCapture`, () => {
   describe(`if(false)`, () => {
     const e = { touches: [{}] }
 
@@ -138,12 +138,12 @@ describe(`onTouchMove`, () => {
       it(`fromTop`, () => test(seedFromTop))
 
       function test(seed) {
-        const active = sinon.stub().returns(false)
+        const canMove = sinon.stub().returns(false)
         const notScroll = sinon.stub().returns(true)
-        const react = { pre: { active, notScroll } }
-        const { onTouchMove } = seed(react)
-        onTouchMove(e)
-        assert.ok(active.calledOnce)
+        const react = { pre: { canMove, notScroll } }
+        const { onTouchMoveCapture } = seed(react)
+        onTouchMoveCapture(e)
+        assert.ok(canMove.calledOnce)
         assert.ok(notScroll.notCalled)
       }
     })
@@ -153,12 +153,12 @@ describe(`onTouchMove`, () => {
       it(`fromRight`, () => test(seedFromRight))
 
       function test(seed) {
-        const active = sinon.stub().returns(true)
+        const canMove = sinon.stub().returns(true)
         const notScroll = sinon.stub().returns(false)
-        const react = { pre: { active, notScroll } }
-        const { onTouchMove } = seed(react)
-        const result = onTouchMove(e)
-        assert.ok(active.calledOnce)
+        const react = { pre: { canMove, notScroll } }
+        const { onTouchMoveCapture } = seed(react)
+        const result = onTouchMoveCapture(e)
+        assert.ok(canMove.calledOnce)
         assert.ok(notScroll.calledOnce)
       }
     })
@@ -173,7 +173,7 @@ describe(`onTouchMove`, () => {
       const value = 10
       const React = () => ({
         state: { value },
-        pre: { active: () => true }
+        pre: { canMove: () => true }
       })
 
       it(`fromBottom`, () => {
@@ -190,8 +190,8 @@ describe(`onTouchMove`, () => {
 
         react.setState = sinon.spy()
 
-        const { onTouchMove } = seedFromBottom(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromBottom(react)
+        onTouchMoveCapture(e)
         assert.ok(react.setState.calledOnce)
       })
 
@@ -210,8 +210,8 @@ describe(`onTouchMove`, () => {
 
         react.setState = sinon.spy()
 
-        const { onTouchMove } = seedFromLeft(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromLeft(react)
+        onTouchMoveCapture(e)
         assert.ok(react.setState.calledOnce)
       })
 
@@ -230,8 +230,8 @@ describe(`onTouchMove`, () => {
 
         react.setState = sinon.spy()
 
-        const { onTouchMove } = seedFromRight(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromRight(react)
+        onTouchMoveCapture(e)
         assert.ok(react.setState.calledOnce)
       })
 
@@ -249,8 +249,8 @@ describe(`onTouchMove`, () => {
 
         react.setState = sinon.spy()
 
-        const { onTouchMove } = seedFromTop(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromTop(react)
+        onTouchMoveCapture(e)
         assert.ok(react.setState.calledOnce)
       })
     })
@@ -263,7 +263,7 @@ describe(`onTouchMove`, () => {
       const value = 0
       const React = () => ({
         state: { value },
-        pre: { active: () => true }
+        pre: { canMove: () => true }
       })
 
       it(`fromBottom`, () => {
@@ -274,8 +274,8 @@ describe(`onTouchMove`, () => {
         react.pre.setSettle = () => {}
         react.pre.setNow = () => {}
 
-        const { onTouchMove } = seedFromBottom(react)
-        const result = onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromBottom(react)
+        const result = onTouchMoveCapture(e)
         assert.equal(result, false)
       })
 
@@ -288,8 +288,8 @@ describe(`onTouchMove`, () => {
         react.pre.setSettle = () => {}
         react.pre.setNow = () => {}
 
-        const { onTouchMove } = seedFromLeft(react)
-        const result = onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromLeft(react)
+        const result = onTouchMoveCapture(e)
         assert.equal(result, false)
       })
 
@@ -302,8 +302,8 @@ describe(`onTouchMove`, () => {
         react.pre.setSettle = () => {}
         react.pre.setNow = () => {}
 
-        const { onTouchMove } = seedFromRight(react)
-        const result = onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromRight(react)
+        const result = onTouchMoveCapture(e)
         assert.equal(result, false)
       })
 
@@ -315,8 +315,8 @@ describe(`onTouchMove`, () => {
         react.pre.setSettle = () => {}
         react.pre.setNow = () => {}
 
-        const { onTouchMove } = seedFromTop(react)
-        const result = onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromTop(react)
+        const result = onTouchMoveCapture(e)
         assert.equal(result, false)
       })
     })
@@ -329,7 +329,7 @@ describe(`onTouchMove`, () => {
       const value = 1
       const React = () => ({
         state: { value },
-        pre: { active: () => true }
+        pre: { canMove: () => true }
       })
 
       it(`fromBottom`, () => {
@@ -341,8 +341,8 @@ describe(`onTouchMove`, () => {
         react.pre.setNow = () => {}
 
         react.come = sinon.spy()
-        const { onTouchMove } = seedFromBottom(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromBottom(react)
+        onTouchMoveCapture(e)
         assert.ok(react.come.calledOnce)
       })
 
@@ -356,8 +356,8 @@ describe(`onTouchMove`, () => {
         react.pre.setNow = () => {}
 
         react.come = sinon.spy()
-        const { onTouchMove } = seedFromLeft(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromLeft(react)
+        onTouchMoveCapture(e)
         assert.ok(react.come.calledOnce)
       })
 
@@ -371,8 +371,8 @@ describe(`onTouchMove`, () => {
         react.pre.setNow = () => {}
 
         react.come = sinon.spy()
-        const { onTouchMove } = seedFromRight(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromRight(react)
+        onTouchMoveCapture(e)
         assert.ok(react.come.calledOnce)
       })
 
@@ -385,18 +385,18 @@ describe(`onTouchMove`, () => {
         react.pre.setNow = () => {}
 
         react.come = sinon.spy()
-        const { onTouchMove } = seedFromTop(react)
-        onTouchMove(e)
+        const { onTouchMoveCapture } = seedFromTop(react)
+        onTouchMoveCapture(e)
         assert.ok(react.come.calledOnce)
       })
     })
   })
 })
 
-describe(`onTouchEnd`, () => {
+describe(`onTouchEndCapture`, () => {
   const Pre = require('../src/Pre.js').default
 
-  describe(`(!pre.active())`, () => {
+  describe(`(!pre.canEnd())`, () => {
     it(`fromBottom`, () => test(seedFromBottom))
     it(`fromLeft`, () => test(seedFromLeft))
     it(`fromRight`, () => test(seedFromRight))
@@ -404,10 +404,10 @@ describe(`onTouchEnd`, () => {
 
     function test(seed) {
       const pre = new Pre()
-      sinon.stub(pre, 'active').returns(false)
-      const { onTouchEnd } = seed({ pre })
-      assert.equal(onTouchEnd(), undefined)
-      assert.ok(pre.active.calledOnce)
+      sinon.stub(pre, 'canEnd').returns(false)
+      const { onTouchEndCapture } = seed({ pre })
+      assert.equal(onTouchEndCapture(), undefined)
+      assert.ok(pre.canEnd.calledOnce)
     }
   })
 
@@ -419,7 +419,7 @@ describe(`onTouchEnd`, () => {
 
     function test(seed) {
       const pre = new Pre()
-      sinon.stub(pre, 'active').returns(true)
+      sinon.stub(pre, 'canEnd').returns(true)
       sinon.stub(pre, 'getNow').returns(Date.now())
       sinon.spy(pre, 'kill')
 
@@ -428,9 +428,14 @@ describe(`onTouchEnd`, () => {
       pre.setSettle(settle)
 
       const react = { pre }
-      const { onTouchEnd } = seed(react)
+      const { onTouchEndCapture } = seed(react)
 
-      onTouchEnd()
+      const stopPropagation = sinon.spy()
+      const preventDefault = sinon.spy()
+      onTouchEndCapture({ stopPropagation, preventDefault })
+
+      assert.ok(stopPropagation.calledOnce)
+      assert.ok(preventDefault.calledOnce)
 
       assert.ok(settle.calledOnce)
       assert.ok(pre.kill.calledOnce)
@@ -445,7 +450,7 @@ describe(`onTouchEnd`, () => {
 
     function test(seed, value) {
       const pre = new Pre()
-      sinon.stub(pre, 'active').returns(true)
+      sinon.stub(pre, 'canEnd').returns(true)
       sinon.stub(pre, 'getSettle').returns(false)
       sinon.spy(pre, 'kill')
 
@@ -458,9 +463,14 @@ describe(`onTouchEnd`, () => {
         quit: sinon.spy()
       }
 
-      const { onTouchEnd } = seed(react)
+      const { onTouchEndCapture } = seed(react)
 
-      onTouchEnd()
+      const stopPropagation = sinon.spy()
+      const preventDefault = sinon.spy()
+      onTouchEndCapture({ stopPropagation, preventDefault })
+
+      assert.ok(stopPropagation.calledOnce)
+      assert.ok(preventDefault.calledOnce)
 
       assert.ok(pre.getSettle.calledOnce)
       assert.ok(react.quit.calledOnce)
@@ -476,7 +486,7 @@ describe(`onTouchEnd`, () => {
 
     function test(seed, value) {
       const pre = new Pre()
-      sinon.stub(pre, 'active').returns(true)
+      sinon.stub(pre, 'canEnd').returns(true)
       sinon.stub(pre, 'getSettle').returns(false)
       sinon.spy(pre, 'kill')
 
@@ -489,9 +499,14 @@ describe(`onTouchEnd`, () => {
         come: sinon.spy()
       }
 
-      const { onTouchEnd } = seed(react)
+      const { onTouchEndCapture } = seed(react)
 
-      onTouchEnd()
+      const stopPropagation = sinon.spy()
+      const preventDefault = sinon.spy()
+      onTouchEndCapture({ stopPropagation, preventDefault })
+
+      assert.ok(stopPropagation.calledOnce)
+      assert.ok(preventDefault.calledOnce)
 
       assert.ok(pre.getSettle.calledOnce)
       assert.ok(react.come.calledOnce)
